Rename the files based on PEP 8
