+++ rjdl/models/album.py
@@ -1,7 +1,7 @@
 import ast
 import requests
 from bs4 import BeautifulSoup
-from .objMusic import Music
+from .music import Music
 
 
 class Album:
